worked on prop validations error and app structure
