+++ src/App.css
@@ -111,3 +111,4 @@ box-shadow: 0 2px 6px rgba(0,0,0,0.08);
   background: linear-gradient(#147EFF, #0F67DA);
   color: white;
 }
+
