+++ src/reducers/actions.js
@@ -1,0 +1,2 @@
+const FILTER_TAGS = (filter) =>  ({type: 'FILTER_TAGS', filters: filter });
+export default FILTER_TAGS;
